requirements.txt
langchain
langchain-core
langchain-text-splitters
langchain-experimental
langchain-openai
langchain-community
tiktoken
langgraph
deepgram-sdk
python-dotenv
pydantic
openai
numpy
moviepy
httpx
rich
vapi-python
anthropic
pypdf
pdf2image